heritrix3_exporter/requirements.txt
hapy-heritrix==0.1.1
prometheus-client==0.0.21
